// src/app/modules/mylist/interfaces/favurl-dto-model.ts
export interface FavurlDto {
  show: boolean;
  deleteConfirmBox: boolean;

  id: number;

  fromid: number;

  toid: number;

  url: string;

  status: number;

  fstatus: number;

  sendtime: Date;

  serial: number;

  readtime: Date;

  comment: string;

  fav: boolean;

  channel: string;

  title: string;

  icon: string;

  host: string;

  nickname: string;

  avatarURL: string;

  urlid: number;

  share: number;

  favs: number;
}